+++ packages/myst-frontmatter/src/thebe/thebe.yml
@@ -125,3 +125,99 @@ cases:
           provider: 'custom'
           repo: 'doi:/10292.131k2'
           ref: ''
+  - title: thebe as empty object is removed
+    raw:
+      thebe: {}
+    normalized: {}
+  - title: thebe strips extra keys
+    raw:
+      thebe:
+        extra: ''
+        server:
+          url: https://my.server.com
+          token: test-secret
+    normalized:
+      thebe:
+        server:
+          url: https://my.server.com
+          token: test-secret
+    warnings: 1
+  - title: custom provider accepts url as repo value
+    raw:
+      thebe:
+        binder:
+          url: https://binder.curvenote.com/services/binder/
+          repo: https://curvenote.com/sub/bundle.zip
+          provider: custom
+    normalized:
+      thebe:
+        binder:
+          url: https://binder.curvenote.com/services/binder/
+          repo: https://curvenote.com/sub/bundle.zip
+          provider: custom
+          ref: ''
+  - title: errors if no repo with custom provider
+    raw:
+      thebe:
+        binder:
+          provider: custom
+    normalized:
+      thebe:
+        binder:
+          provider: custom
+          ref: ''
+          url: https://mybinder.org/
+    errors: 2
+  - title: 'thebe: "server" is not valid'
+    raw:
+      thebe: server
+    normalized: {}
+    errors: 1
+  - title: thebe.server must be an object
+    raw:
+      thebe:
+        server: true
+    normalized: {}
+    errors: 1
+  - title: thebe.server must have url and token fields - empty
+    raw:
+      thebe:
+        server: {}
+    normalized: {}
+    errors: 1
+  - title: thebe.server must have url and token fields - no url
+    raw:
+      thebe:
+        server:
+          token: my-secret-secret
+    normalized: {}
+    errors: 1
+  - title: thebe.server must have url and token fields - not a url string
+    raw:
+      thebe:
+        server:
+          url: not-a-url
+          token: my-secret-secret
+    normalized: {}
+    errors: 1
+  - title: thebe.server must have url and token fields - no token
+    raw:
+      thebe:
+        server:
+          url: http://localhost:90090
+    normalized: {}
+    errors: 1
+  - title: thebe.binder - extra fields wil be stripped
+    raw:
+      thebe:
+        binder:
+          url: http://localhost:9090
+          extra: field
+    normalized:
+      thebe:
+        binder:
+          url: http://localhost:9090
+          repo: executablebooks/thebe-binder-base
+          ref: HEAD
+          provider: github
+    warnings: 1
