+++ packages/myst-frontmatter/tests/examples.spec.ts
@@ -67,9 +67,6 @@ casesList.forEach(({ title, frontmatter, cases }) => {
           console.log(opts.messages.errors);
         }
         expect(result).toEqual(normalized);
-        if ((opts.messages.warnings?.length ?? 0) !== (warnings ?? 0)) {
-          console.log(opts.messages.warnings);
-        }
         expect(opts.messages.warnings?.length ?? 0).toBe(warnings ?? 0);
         expect(opts.messages.errors?.length ?? 0).toBe(errors ?? 0);
